pyproject.toml
[project]
name = "table-reservation-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.0",
]
